Improve changes #3

+++ app/src/main/java/de/koenidv/sph/objects/Change.kt
@@ -22,12 +22,12 @@ data class Change(
         var room: String? = null, // New room, i.e. M115 - Uppercase
         var room_before: String? = null, // Old room. Per SPH implementation as of now, this will always be null - Uppercase
         var description: String? = null, // Description of the event
-        var sortLesson: Float? = null // Only for sql ordering
+        var sortLesson: Double? = null // Only for sql ordering
 ) {
 
     init {
-        sortLesson = if (lessons.size == 1) lessons[0].toFloat()
-        else (lessons[0] + 0.01 * lessons[lessons.size - 1]).toFloat()
+        sortLesson = if (lessons.size == 1) lessons[0].toDouble()
+        else (lessons[0] + 0.01 * lessons[lessons.size - 1])
         if (description == "Entfällt" || description == "Entfällt; Verlegung auf Entfall für Lehrer")
             description = null
     }
